### 2020-08-20 #2587 @whedon
[PRE REVIEW]: PINE: An open source collaborative text annotation tool
**Submitting author:** @brantapl (<a href="http://orcid.org/0000-0002-6174-4201">Brant Chee</a>)
**Repository:** <a href="https://github.com/jhuapl/PINE" target ="_blank">https://github.com/jhuapl/PINE</a>
**Version:** v1.0
