### 2020-09-09 e8161dc
Update with latest developments.
Including:
* Updated npm dependencies.
* Improvements for downloading large collections.
* Additions to python client and python client tests.
* Bug fix for collection creation when using eve auth.
