### 2020-09-10 #2587 @majensen
@wdduncan - would you have time and interest to review [this work](#2587) for [JOSS](https://joss.theoj.org)? Consideration is much appreciated!
