### 2020-09-10 #2587 @majensen
@yongqunh - this is @majensen, we have worked in the KPMP consortium together.
As a Topic Editor for [JOSS](https://joss.theoj.org), I am wondering - would you be willing to perform a software review on [this work](#2587) for the journal?
An overview of the review process is [here](https://joss.readthedocs.io/en/latest/reviewer_guidelines.html). Please let me know here in this thread, or feel free to reach out to me at maj \-dot\- fortinbras \-at\- gmail \-dot\- com.
Thanks for considering!
