### 2020-09-15 #2587 @wdduncan
> @wdduncan - would you have time and interest to review [this work](#2587) for [JOSS](https://joss.theoj.org)? Consideration is much appreciated!

@majensen Sorry, I'm going to have to pass on this one. I have too many other reviews in the queue :(
