### 2020-09-16 #2587 @majensen
Hey @tvalerius - I'm a topic editor for the online [Journal of Open Source Software](https://joss.theoj.org), and I was thinking that [this work](#2587) would be right up your alley. Would you be willing to contribute your expertise on a review?
All reviews are open and performed online here in GitHub. [Here is an overview of the process](https://joss.readthedocs.io/en/latest/reviewer_guidelines.html). Please respond in this thread, or feel free to reach out to me at _maj \-dot\- fortinbras \-at\- gmail \-dot\- com_, or my NIH address.

Really appreciate your consideration -- @majensen
