### 2020-09-23 #2587 @majensen
@balhoff - can I trouble you again to consider a [JOSS](https://joss.theoj.org) review of [this work](#2487)? All my other ontology experts seem to be working hard on their ontologies and I haven't got a bite.... Appreciate your consideration! @majensen
