### 2020-09-23 #2587 @rvosa
> @rvosa -- long time! I'm an editor now for the online journal [JOSS](https://joss.theoj.org). Would you be willing to review [this work](#2587)? All reviews are open and performed online here in GitHub. [Here is an overview](https://joss.readthedocs.io/en/latest/reviewer_guidelines.html) of the process. Please respond in this thread, or feel free to reach out to me at maj -dot- fortinbras -at- gmail -dot- com. I really appreciate your consideration --@majensen

Hi @majensen! Nice to hear from you. I'm sorry but I'll have to decline. Things are a bit hectic right now and this project also seems a bit outside of my expertise. Best, Rutger
