### 2020-09-23 #2587 @majensen
Hey @sap218 -- would you have time and inclination to review [this work](#2587) for [JOSS](https://joss.theoj.org)? Would love to have your expertise on this.
Thanks! --@majensen
