### 2020-09-24 #2587 @brantapl
@majensen Thanks for all the help. Do you mean JOSS reviewers or outside of the community? Are their any requirements for reviewers?
