### 2020-09-24 #2587 @majensen
@brantapl can be anyone who has some expertise in the field and is willing to do the software testing. Outside the JOSS community is fine, then we can being them in!
