### 2020-10-06 #2587 @majensen
@callahantiff, @balhoff suggested that you would be a good reviewer for [JOSS](https://joss.theoj.org) on [this work](#2587). Would you be willing to provide your expertise? [Here](https://joss.readthedocs.io/en/latest/reviewer_guidelines.html) is an overview of the reviewing process.
Thanks for your consideration --@majensen
