### 2020-10-06 #2587 @callahantiff
> @callahantiff, @balhoff suggested that you would be a good reviewer for [JOSS](https://joss.theoj.org) on [this work](#2587). Would you be willing to provide your expertise? [Here](https://joss.readthedocs.io/en/latest/reviewer_guidelines.html) is an overview of the reviewing process.
> Thanks for your consideration --@majensen

Hi @majensen - Thanks for the invitation to review (thanks for recommending me @balhoff). I feel horrible to decline given that it looks like it's been hard to find reviewers, but I am also currently slammed (have 5 reviews due this week) ☹️

I do have a few other people in mind that I think would be very qualified to review this work: @bill-baumgartner, @jmbanda, @mboguslav, @KevinBretonnelCohen, or @jdkim.

Thank you again for the invitation, please keep me in mind for future papers!
