### 2020-10-06 #2587 @majensen
Thanks @callahantiff for your response and suggestions! Any of those mentioned, can you ping back on this thread if you are willing to review?
