### 2020-10-07 #2726 @majensen
@bill-baumgartner @bensonml - your TL;DR:
* Your review checklists are at the top of the issue
* PDF of draft paper is [here](https://raw.githubusercontent.com/openjournals/joss-papers/joss.02726/joss.02726/10.21105.joss.02726.pdf)
* The software is in [this repo](https://github.com/jhuapl/PINE)
* Make comments for improvements in this issue. It is a conversation with the author; goal is to improve the package in "real time"
* For more technical issues, make an issue directly in the repo.
* Have fun
* Ping me with any questions
Thank you for your service!
