+++ frontend/annotation/src/app/component/collection-details/collection-details.component.ts
@@ -1,9 +1,11 @@
 /*(C) 2019 The Johns Hopkins University Applied Physics Laboratory LLC. */
 import { Component, OnInit, ViewChild, AfterViewInit, Inject } from "@angular/core";
 import { MatPaginator, MatTableDataSource, MatSort, MatDialog, MatDialogRef, MAT_DIALOG_DATA } from "@angular/material";
-import { ActivatedRoute } from "@angular/router";
+import { ActivatedRoute, Router } from "@angular/router";
 import { HttpErrorResponse } from "@angular/common/http";
 
+import * as _ from "lodash";
+import * as moment from "moment";
 import { take } from "rxjs/operators";
 
 import { PATHS, PARAMS } from "../../app.paths";
@@ -17,28 +19,29 @@ import { EventService } from "../../service/event/event.service";
 import { PipelineService } from "../../service/pipeline/pipeline.service";
 import { MetricsService } from "../../service/metrics/metrics.service";
 
-import { Annotation } from  "../../model/annotation";
+import { Annotation } from "../../model/annotation";
 import { Document } from "../../model/document";
 import { Classifier } from "../../model/classifier";
 import { Collection, DownloadCollectionData, METADATA_TITLE } from "../../model/collection";
 import { Pipeline } from "../../model/pipeline";
-import {DocumentMenuItem} from "../nav-collection-menu/nav-collection-menu.component";
+import { DocumentMenuItem } from "../nav-collection-menu/nav-collection-menu.component";
 import { Metric } from "../../model/metrics";
 import { IaaReportingService } from '../../service/iaa-reporting/iaa-reporting.service';
 import { IAAReport } from '../../model/iaareport';
 import { DownloadCollectionDataDialogComponent } from '../download-collection-data.dialog/download-collection-data.dialog.component';
 import { ImageCollectionUploaderComponent, dialog } from "../image-collection-uploader/image-collection-uploader.component";
+import { AddDocumentComponent, AddDocumentDialogData } from '../add-document/add-document.component';
 
 export interface AnnotatorDialogData {
-  annotator: string;
+    annotator: string;
 }
 
 export interface ViewerDialogData {
-  viewer: string;
+    viewer: string;
 }
 
 export interface LabelDialogData {
-  label: string;
+    label: string;
 }
 
 export interface DocumentRow {
@@ -47,7 +50,8 @@ export interface DocumentRow {
     last_updated: Date;
     text_start: string;
     annotated: boolean;
-    ann_agreement: number
+    ann_agreement: number;
+    created: Date;
 }
 
 @Component({
@@ -61,6 +65,8 @@ export class CollectionDetailsComponent implements OnInit {
 
     public readonly PATHS = PATHS;
 
+    public tabIndex: number = 0;
+
     public loading = false;
     public canArchive = false;
 
@@ -70,9 +76,10 @@ export class CollectionDetailsComponent implements OnInit {
     public pipeline: Pipeline;
     documents: DocumentRow[];
     dataSource = new MatTableDataSource<DocumentRow>();
+    oldestDocId: string = null;
     nextDocId: string = null;
     metrics: object;
-    iaa_report : IAAReport;
+    iaa_report: IAAReport;
     private new_annotator: string = null;
     private new_viewer: string = null;
     private new_label: string = null;
@@ -84,22 +91,30 @@ export class CollectionDetailsComponent implements OnInit {
     @ViewChild(MatPaginator) paginator: MatPaginator;
     @ViewChild(MatSort) sort: MatSort;
 
-    constructor(private route: ActivatedRoute,
-                private annotationService: AnnotationService,
-                private collectionService: CollectionRepositoryService,
-                private documentsService: DocumentRepositoryService,
-                private pipelineService: PipelineService,
-                private metricsService: MetricsService,
[…]
