+++ frontend/annotation/src/app/component/toolbar/index.ts
@@ -1,0 +1,1 @@
+export * from './toolbar/toolbar.component';
