### 2020-10-09 #2726 @majensen
@openjournals/dev - [pdf](https://github.com/openjournals/joss-papers/blob/joss.02726/joss.02726/10.21105.joss.02726.pdf) generated on a [custom branch](https://github.com/JHUAPL/PINE/tree/JOSS) is missing the rendering of figure "architecture.png". That file appears to be present in the directory containing paper.md in the branch. Can you advise? Thanks
