### 2020-10-19 JHUAPL/PINE#10 @bensonml
Fail during `run_docker_compose.sh --build`
I'm trying to install and run PINE using Docker, following `README.md` instructions - but it is failing.

System: macOS Catalina (10.15.6), working repo at commit `ebfc2c6`

Steps/What I've done:
1. Ran `pipelines/download_resources.sh` script (components were downloaded and unzipped as expected)
2. Set VEGAS_CLIENT_SECRET in `.env`
3. Set PINE_VERSION via `version.sh`
4. Failed build during `./run_docker_compose.sh --build` with error:
```
...
Step 32/33 : RUN npm run prod
 ---> Running in 4f1f4bddc645

> pine@1.1.0 prod /nlp-web-app/frontend
> ng build --prod

Date: 2020-10-19T20:44:05.844Z
Hash: d849bbb19b5a5d0bde38
Time: 23000ms
chunk {0} runtime.80ab492fe3d778817936.js (runtime) 1.41 kB [entry] [rendered]
chunk {1} main.202765e82fd65fa4f907.js (main) 128 bytes [initial] [rendered]
chunk {2} polyfills.cccb2ee14eb4f9d80fbd.js (polyfills) 130 bytes [initial] [rendered]
chunk {3} styles.f719768ddbd0ac85d51d.css (styles) 79.2 kB [initial] [rendered]
chunk {scripts} scripts.bbe10e1331279f9903a7.js (scripts) 87.1 kB [entry] [rendered]

ERROR in : Cannot determine the module for class HomeComponent in /nlp-web-app/frontend/src/app/component/home/home.component.ts! Add HomeComponent to the NgModule to fix it.

npm ERR! code ELIFECYCLE
npm ERR! errno 1
npm ERR! pine@1.1.0 prod: `ng build --prod`
npm ERR! Exit status 1
npm ERR!
npm ERR! Failed at the pine@1.1.0 prod script.
npm ERR! This is probably not a problem with npm. There is likely additional logging output above.

npm ERR! A complete log of this run can be found in:
npm ERR!     /root/.npm/_logs/2020-10-19T20_44_05_978Z-debug.log
ERROR: Service 'frontend_annotation' failed to build : The command '/bin/sh -c npm run prod' returned a non-zero code: 1
```

attaching output from `script` command (typescript of terminal session)
[pine-install-docker-typescript.log](https://github.com/JHUAPL/PINE/files/5404665/pine-install-docker-typescript.log)
