### 2020-11-02 #2726 @majensen
@brantapl can you review the above issues https://github.com/JHUAPL/PINE/pull/8 https://github.com/JHUAPL/PINE/issues/10 ?  These appear to be blockers for the reviews

I know this is a difficult season for many of us, so just let me know what your timeframe is.
