### 2020-11-03 #2726 @bwkchee
@all

> @brantapl can you review the above issues [JHUAPL/PINE#8](https://github.com/JHUAPL/PINE/pull/8) [JHUAPL/PINE#10](https://github.com/JHUAPL/PINE/issues/10) ? These appear to be blockers for the reviews
>
> I know this is a difficult season for many of us, so just let me know what your timeframe is.

Looking into this now.  It will probably be a couple days since people are out today (for the election).  Thank you!
