### 2020-11-04 a939fb4
Fix bugs affecting docker/prod build and tests.
