+++ frontend/annotation/src/app/component/nav-collection-menu/nav-collection-menu.component.spec.ts
@@ -1,27 +1,0 @@
-/*(C) 2019 The Johns Hopkins University Applied Physics Laboratory LLC. */
-
-import { async, ComponentFixture, TestBed } from '@angular/core/testing';
-
-import { NavCollectionMenuComponent } from './nav-collection-menu.component';
-
-describe('NavCollectionMenuComponent', () => {
-  let component: NavCollectionMenuComponent;
-  let fixture: ComponentFixture<NavCollectionMenuComponent>;
-
-  beforeEach(async(() => {
-    TestBed.configureTestingModule({
-      declarations: [ NavCollectionMenuComponent ]
-    })
-    .compileComponents();
-  }));
-
-  beforeEach(() => {
-    fixture = TestBed.createComponent(NavCollectionMenuComponent);
-    component = fixture.componentInstance;
-    fixture.detectChanges();
-  });
-
-  it('should create', () => {
-    expect(component).toBeTruthy();
-  });
-});
