### 2020-11-04 9267fb7
Remove code that was mistakenly left in public branch.
