### 2020-11-04 JHUAPL/PINE#10 @jhuapl-lglenden
This particular issue was caused by missing some files that were deleted in our internal repo but not in the github repo.  It has been fixed with #12 .  However, there are some other build issues that we are working on resolving before the docker build can work.
