### 2020-11-06 JHUAPL/PINE#10 @jhuapl-lglenden
The docker build should now build without issue, from #13
