### 2020-11-10 #2726 @bill-baumgartner
> @bill-baumgartner have you made any headway on this review? Let me know if you are having any issues-- thank you!

@majensen I have made some progress, but am having install issues as noted in Issue [#14](https://github.com/JHUAPL/PINE/issues/14).

One question for you: what is the preferred method for making comments on the manuscript? Should we simply add comments on this review page?
