### 2020-11-11 JHUAPL/PINE#14 @bwkchee
Thank you for the detailed bug report.  We're currently exploring our options, it is due to an update in the mongo package.  Hope to have this resolved shortly.
