### 2020-11-11 #2726 @majensen
@bill-baumgartner thanks for the update - @brantapl can you or @bwkchee look into the issue noted above?

Feel free to make any paper comments right here. If there are technical issues that need a detailed discussion, you can create an issue in the repo and refer to that here in general comments.
thanks
