### 2020-11-13 JHUAPL/PINE#14 @jhuapl-lglenden
This should be resolved by #15 which updates the MongoDB package being used.
