### 2020-11-17 JHUAPL/PINE#10 @bensonml
Yes - confirmed it builds and runs (built on Ubuntu 20.04 and macOS Catalina 10.15.7 )
very sorry for the late reply; I had a glitch building on Ubuntu and I was running the problem down, but it was something on my end, and it works)
