### 2020-11-17 #2726 @bill-baumgartner
Here are my initial comments on the paper:

- `PMAP` is used in the first sentence but is not defined until later in the document.
- Although NER and document classification are mentioned in the paper, the types of annotation supported by PINE are not explicitly enumerated. A sentence or two in the introduction listing the supported types of annotation would be helpful to the reader.
- Inter-annotator agreement is mentioned however the specific metric used is not detailed in the manuscript as far as I could tell.
- You may want to include the [Inception annotation platform](https://inception-project.github.io/) in the `Related work` section -- I believe it has an active learning component so it seems relevant.
- A spell checker is needed for some typos throughout the manuscript.
- The description of the `Pipeline` wrapper on page 3 might be considered API documentation, and therefore may belong in the software documentation instead of the paper according to the [review criteria](https://joss.readthedocs.io/en/latest/review_criteria.html).
- The manuscript would benefit from one or more figures displaying the annotation UI in my opinion.
- References that are just URLs (e.g. Gunicorn, Eve) could be replaced by footnotes if that is permitted.
- The Dublin Core reference has numbers where the author is typically listed.

Hope those are helpful. Please let me know if something is unclear.
