### 2020-11-18 JHUAPL/PINE#17 @bill-baumgartner
Error when creating new admin user
Hi, An error is observed when trying to create a new administrator.

System: `macOS Catalina 10.15.6`
Docker: `Docker version 19.03.13, build 4484c46d9d`; `docker-compose version 1.27.4, build 40524192`
Browser: `Google Chrome (Version 86.0.4240.198)`
Repo commit: 69312cf

Steps taken:

- `./pipelines/download_resources.sh`
- changed `AUTH_MODULE=vegas` to `AUTH_MODULE=eve` in `.env`
- `export PINE_VERSION=$(./version.sh)`
- `docker-compose build`
- `docker-compose up`
- `docker-compose exec backend scripts/data/add_admin.sh admin 1234`

```
> docker-compose exec backend scripts/data/add_admin.sh admin 1234
2020-11-18 02:47:44 [INFO] pine.backend.log: Set logging configuration from file /nlp-web-app/shared/logging.python.dev.json
2020-11-18 02:47:44 [INFO] pine.backend.job_manager.service: Using redis host redis:6379
Traceback (most recent call last):
  File "/usr/local/bin/flask", line 8, in <module>
    sys.exit(main())
  File "/usr/local/lib/python3.6/dist-packages/flask/cli.py", line 967, in main
    cli.main(args=sys.argv[1:], prog_name="python -m flask" if as_module else None)
  File "/usr/local/lib/python3.6/dist-packages/flask/cli.py", line 586, in main
    return super(FlaskGroup, self).main(*args, **kwargs)
  File "/usr/local/lib/python3.6/dist-packages/click/core.py", line 782, in main
    rv = self.invoke(ctx)
  File "/usr/local/lib/python3.6/dist-packages/click/core.py", line 1259, in invoke
    return _process_result(sub_ctx.command.invoke(sub_ctx))
  File "/usr/local/lib/python3.6/dist-packages/click/core.py", line 1066, in invoke
    return ctx.invoke(self.callback, **ctx.params)
  File "/usr/local/lib/python3.6/dist-packages/click/core.py", line 610, in invoke
    return callback(*args, **kwargs)
  File "/usr/local/lib/python3.6/dist-packages/click/decorators.py", line 21, in new_func
    return f(get_current_context(), *args, **kwargs)
  File "/usr/local/lib/python3.6/dist-packages/flask/cli.py", line 426, in decorator
    return __ctx.invoke(f, *args, **kwargs)
  File "/usr/local/lib/python3.6/dist-packages/click/core.py", line 610, in invoke
    return callback(*args, **kwargs)
  File "/nlp-web-app/backend/pine/backend/data/users.py", line 78, in add_admin_command
    abort(resp.status_code)
  File "/usr/local/lib/python3.6/dist-packages/werkzeug/exceptions.py", line 822, in abort
    return _aborter(status, *args, **kwargs)
  File "/usr/local/lib/python3.6/dist-packages/werkzeug/exceptions.py", line 807, in __call__
    raise self.mapping[code](*args, **kwargs)
werkzeug.exceptions.UnprocessableEntity: 422 Unprocessable Entity: The request was well-formed but was unable to be followed due to semantic errors.
```
