### 2020-11-20 #2726 @bensonml
## Comments on Paper:
I concur with the above comments about the paper.

Here are my own comments:

1. **Grammar**: add "the"
   1. We have developed an extensible framework to support annotation of plain text
   2. We have developed an extensible framework to support ***the*** annotation of plain text

2. **References: Citation Syntax**: maybe use a reference or inline url instead of a footnote for PMAP?  Similar journals such as PLOS recommends either for referencing a website:
   1. reference style:  PMAP: The Johns Hopkins Precision Medicine Analytics Platform. 2019. *Precision Medicine Portal | Johns Hopkins Medicine*. [online] Available at: <https://pm.jh.edu> [Accessed 19 November 2020].
   2. inline style: "...on the Johns Hopkins University Precision Medicine Analytics Platform (PMAP) (https://pm.jh.edu)"

3. **Grammar**: better wording ?
   from: "Docker Compose and Kubernetes allow for horizontal scaling of various components if necessary such as machine learning containers, annotation user interface containers, etc. "
   to: "Docker Compose and Kubernetes allow for horizontal scaling of various components **~~if necessary~~** such as machine learning containers, annotation user interface containers, etc. "

4. **References: Citation Syntax**:  the  Dublin Core Metadata Elements Set reference needs to include the initialism "ISO" instead of just the number "15836". Examples of reference syntax for the Dublin Core set in literature:
   1. Dublin Core (ISO 15836) http://dublincore.org/
      1. as used in https://www.ncbi.nlm.nih.gov/pmc/articles/PMC3224535/#B61
   2. The Dublin Core Metadata Initiative (DCMI) Element Set (ISO 15836:2009)
      1. as used in https://pubmed.ncbi.nlm.nih.gov/22075810/
   3. ISO Standard 15836:2009 of February 2009 [ISO15836]
      1. as described on https://www.dublincore.org/specifications/dublin-core/dces/

5. **Grammar**:
    from : PINE is a web based application that
   to:  PINE is a web**-**based application that

6. **Grammar**:
    from: changes are tracked and store. This
   to: changes are tracked and store**d**. This

7. **References: Citation Syntax** for: Gunicorn Python, W.  Don't think references are really needed for frameworks (e.g. EVE) or technologies like Mongo, Redis, etc.  I looked at several other papers published in JOSS, and they do not have such references.  Concur that maybe they could be either footnotes, or simply inline citations (just put the url in the text).  Besides, the reference for Gunicorn just doesn't make sense:  "Python, W."

8. **Glitch**: The figure **??** depicts thearchitecture of PINE.
   1. This was supposedly fixed as part of https://github.com/JHUAPL/PINE/pull/8 , but I'm not able to see it. I don't know if it is possible to re-build/render the pdf, but this needs to be verified before being published.

9. **Software Documentation including API**: the discussion about need to implement the Pipeline class appears to fall under the material that should not be included in the paper, as described under the review criteria in the box ([link to review criteria](https://joss.readthedocs.io/en/latest/review_criteria.html#review-items)). It is recommended to rewrite this section. I believe it is appropriate to point out this is the place for extending/adding pipelines, but doesn't necessarily need to list all of the functions in this class. Furthermore, I could not find this information is not found anywhere in the PINE github repo. (in other words: this needs to be part of documentation in the PINE github repo)
   1. "Note the paper *should not* include software documentation such  as API (Application Programming Interface) functionality, as this should be outlined in the software documentation."

10. **References: Citation Syntax and grammar**: The Acronyms section
    1. Precision Medicine Analytics **Platoform**  --> Precision Medicine Analytics Platform
    2. Use semicolons to separate acronyms. It currently comes across as a single run on sentence:
       1. PMAP - Precision Medicine Analytics Platoform NLP - Natural Language processing PINE -PMAP Interface for NLP Experimentation
       2. PMAP - Precision Medicine Analytics Platform; NLP - Natural Language processing; PINE -PMAP Interface for NLP Experimentation;

11. **Grammar**:
    from: The **wraper** is similar
    to: The wrapper is similar

12. **Grammar:**
    from: Turnbull, J. (2014). The **docker book**: Containerization is the new virtualization.
    to: Turnbull, J. (2014). The Docker Book: Containerization is the new virtualization.

--------
## Functionality
**Installation:** Does installation proceed as outlined in the documentation?

1. "Docker Environments" installation/running ok: :white_check_mark:
2. "Development Environment" installation/running :warning:
   1. I needed to first run `pip install -U setuptools` in order to get the backend to work
   2. just `./setup_dev_stack.sh` did NOT work, nor going into backend/ and starting there (following README.md instructions).
   3. I was testing on a clean ubuntu 20.04 install running python 3.6; It would stop with essentially the same error (see below) for two packages:
* overrides==2.8.0
* python-json-logger==0.1.11

but the error would be resolved if I ran `pip install -U setuptools` before `pipenv install --dev`

key to figuring this out was https://stackoverflow.com/questions/36296134/attributeerror-install-layout-when-attempting-to-install-a-package-in-a-virtual

error snippet:
```
      File "/home/mark/.pyenv/versions/3.6.0/lib/python3.6/distutils/cmd.py", line 103, in __getattr__
        raise AttributeError(attr)
    AttributeError: install_layout
```
