### 2020-12-21 #2726 @majensen
@brantapl @bwkchee - So it's been about a month since my last ping. Would it be better to pause this review until you have more time to work on the paper?
@bill-baumgartner any new progress?
thanks
