### 2020-12-21 #2726 @bill-baumgartner
@majensen - no new progress on my end. I have been able to install the software, but have not been able to evaluate its use. I believe I am currently blocked by [this issue](https://github.com/JHUAPL/PINE/issues/17).
