### 2021-01-08 523dbf5
Update from latest develop branch.
Including:
* README updates
* Fixes to user auth scripts
