### 2021-01-08 JHUAPL/PINE#17 @jhuapl-lglenden
The issues with these scripts should have been fixed by #19 .
