### 2021-01-14 #2726 @majensen
@jhuapl-lglenden - Can you comment on the status of [this paper](#2726) in JOSS, submitted by your colleague @bwkchee? Thanks for any help you can provide. Feel free to contact me at maj -dot- fortinbras -at- gmail -dot- com.
