### 2021-01-26 82b671f
Updates from internal development branch.
Including:
* Adding pagination to documents table in collection view
* Miscellaneous UI fixes and improvements
