+++ frontend/annotation/src/app/component/collection-details/document-data-source.ts
@@ -1,159 +1,0 @@
-/*(C) 2019 The Johns Hopkins University Applied Physics Laboratory LLC. */
-
-import { ElementRef, OnDestroy } from "@angular/core";
-import { CollectionViewer, DataSource } from "@angular/cdk/collections";
-import { MatPaginator, MatSort } from "@angular/material";
-import { HttpErrorResponse } from "@angular/common/http";
-
-import { Observable, BehaviorSubject, Subscription, forkJoin, merge, fromEvent } from "rxjs";
-import { take, debounceTime, distinctUntilChanged, tap } from "rxjs/operators";
-
-import { Document } from "../../model/document";
-import { IAAReport } from "../../model/iaareport";
-
-import { AuthService } from "../../service/auth/auth.service";
-import { DocumentRepositoryService, PaginatedDocuments } from "../../service/document-repository/document-repository.service";
-import { IaaReportingService } from "../../service/iaa-reporting/iaa-reporting.service";
-
-export interface DocumentRow {
-    id: string;
-    creator: string;
-    last_updated: Date;
-    text_start: string;
-    annotated: boolean;
-    ann_agreement: number;
-}
-
-const fieldMapping: {[column: string]: string} = {
-    "id": "_id",
-    "creator": "creator_id",
-    "last_updated": "_updated",
-    "text_start": "text"
-}
-
-export class DocumentDataSource implements OnDestroy, DataSource<DocumentRow> {
-
-    private documentSubject = new BehaviorSubject<DocumentRow[]>([]);
-    private loadingSubject = new BehaviorSubject<boolean>(false);
-    private subscription: Subscription;
-
-    public loading$ = this.loadingSubject.asObservable();
-    public collectionReport: IAAReport;
-    public length: number;
-
-    public latestError: string = null;
-
-    constructor(private documentService: DocumentRepositoryService,
-                private auth: AuthService,
-                private iaa: IaaReportingService,
-                private collectionId: string = null,
-                public truncate: boolean = true,
-                public truncateLength: number = Document.DEFAULT_PREVIEW_LENGTH) { }
-
-    ngOnDestroy() {
-        if(this.subscription) {
-            this.subscription.unsubscribe();
-            this.subscription = undefined;
-        }
-    }
-
-    connect(_: CollectionViewer): Observable<DocumentRow[]> {
-        return this.documentSubject.asObservable();
-    }
-
-    disconnect(_: CollectionViewer): void {
-        this.documentSubject.complete();
-        this.loadingSubject.complete();
-    }
-
-    public setPaginatorSortAndFilter(paginator: MatPaginator, sort: MatSort, filter: ElementRef) {
-        if(this.subscription) {
-            this.subscription.unsubscribe();
-        }
-        this.loadFrom(paginator, sort, filter);
-        this.subscription =
-            merge(
-                paginator.page,
-                sort.sortChange.pipe(
-                    tap(() => paginator.pageIndex = 0)),
-                fromEvent(filter.nativeElement, "keyup").pipe(
-                    debounceTime(500),
-                    distinctUntilChanged(),
-                    tap(() => paginator.pageIndex = 0))
-            ).subscribe((_) => {
-                this.loadFrom(paginator, sort, filter);
-            }, (error: Error) => {
-                console.error(error);
-                this.latestError = error.message;
-            });
-    }
-
-    public setCollection(collectionId: string): Observable<boolean> {
-        this.loadingSubject.next(true);
-        this.latestError = null;
-        return new Observable<boolean>((observer) => {
-            this.collectionId = collectionId;
-            forkJoin(
-                this.iaa.getIIAReportByCollection(this.collectionId),
[…]
