### 2021-02-05 784deb5
Updates from internal developments.
Including:
* Adding pagination to documents table of collection view.
* Miscellaenous UI bug fixes and improvements.
