+++ frontend/annotation/src/app/component/collection-details/collection-details.component.ts
@@ -1,11 +1,13 @@
-/*(C) 2019 The Johns Hopkins University Applied Physics Laboratory LLC. */
-import { Component, OnInit, ViewChild, Inject } from "@angular/core";
-import { MatPaginator, MatTableDataSource, MatSort, MatDialog, MatDialogRef, MAT_DIALOG_DATA } from "@angular/material";
+/* (C) 2019 The Johns Hopkins University Applied Physics Laboratory LLC. */
+import { Component, OnInit, ViewChild, AfterViewInit, Inject, OnDestroy, ElementRef } from "@angular/core";
+import { MatPaginator, MatSort, MatDialog, MatDialogRef, MAT_DIALOG_DATA } from "@angular/material";
 import { ActivatedRoute, Router } from "@angular/router";
 import { HttpErrorResponse } from "@angular/common/http";
 
+import { forkJoin, BehaviorSubject } from "rxjs";
+import { filter, take } from "rxjs/operators";
+
 import * as _ from "lodash";
-import { take } from "rxjs/operators";
 
 import { PATHS, PARAMS } from "../../app.paths";
 
@@ -17,15 +19,17 @@ import { EventService } from "../../service/event/event.service";
 import { PipelineService } from "../../service/pipeline/pipeline.service";
 import { MetricsService } from "../../service/metrics/metrics.service";
 
-import { Document } from "../../model/document";
 import { Classifier } from "../../model/classifier";
 import { Collection, DownloadCollectionData, METADATA_TITLE } from "../../model/collection";
 import { Pipeline } from "../../model/pipeline";
+import { Metric } from "../../model/metrics";
 import { IaaReportingService } from '../../service/iaa-reporting/iaa-reporting.service';
 import { IAAReport } from '../../model/iaareport';
 import { DownloadCollectionDataDialogComponent } from '../download-collection-data.dialog/download-collection-data.dialog.component';
 import { ImageCollectionUploaderComponent, dialog } from "../image-collection-uploader/image-collection-uploader.component";
 import { AddDocumentComponent, AddDocumentDialogData } from '../add-document/add-document.component';
+
+import { DocumentDataSource } from "./document-data-source";
 
 export interface AnnotatorDialogData {
     annotator: string;
@@ -39,22 +43,12 @@ export interface LabelDialogData {
     label: string;
 }
 
-export interface DocumentRow {
-    id: string;
-    creator: string;
-    last_updated: Date;
-    text_start: string;
-    annotated: boolean;
-    ann_agreement: number;
-    created: Date;
-}
-
 @Component({
     selector: "app-collection-details",
     templateUrl: "./collection-details.component.html",
     styleUrls: ["./collection-details.component.css"]
 })
-export class CollectionDetailsComponent implements OnInit {
+export class CollectionDetailsComponent implements OnInit, AfterViewInit, OnDestroy {
 
     public static readonly SUBTITLE = "Collection Details";
 
@@ -69,9 +63,7 @@ export class CollectionDetailsComponent implements OnInit {
     collection: Collection;
     public classifier: Classifier;
     public pipeline: Pipeline;
-    documents: DocumentRow[];
-    dataSource = new MatTableDataSource<DocumentRow>();
-    oldestDocId: string = null;
+    public documents: DocumentDataSource;
     nextDocId: string = null;
     metrics: object;
     iaa_report: IAAReport;
@@ -82,9 +74,10 @@ export class CollectionDetailsComponent implements OnInit {
     public can_add_documents: boolean = false;
     public can_add_images: boolean = false;
 
-
-    @ViewChild(MatPaginator) paginator: MatPaginator;
-    @ViewChild(MatSort) sort: MatSort;
[…]
