### 2021-02-28 #2726 @majensen
@danielskatz it has been 4 weeks since the in-danger-of-rejection label was added. See https://github.com/openjournals/joss-reviews/issues/2726#issuecomment-770249692.
